[cli/feature]: split feature command into a separate file (#1034)
create new feature.py for both config and show, move the code into feature.py
move common code into sonic-utiltities.common.cli.py
